tests/wpt/web-platform-tests/web-nfc/NFCReadingEvent_constructor.https.html

<!DOCTYPE html>
<title>NFCReadingEvent constructor</title>
<link rel="help" href="https://w3c.github.io/web-nfc/#dom-nfcreadingevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/nfc_help.js"></script>
<script>

  test(() => {
    assert_equals(NFCReadingEvent.length, 2);
    assert_throws(new TypeError, () => new NFCReadingEvent('message'));
  }, 'NFCReadingEvent constructor without init dict');

  test(() => {
    const message = createMessage([createJsonRecord(test_json_data)]);
    const event = new NFCReadingEvent('type', {serialNumber: '', message: message});
    assert_equals(event.type, 'type', 'type');
    assert_equals(event.serialNumber, '', 'serialNumber');
    assertWebNDEFMessagesEqual(event.message, message, 'message');
  }, 'NFCReadingEvent constructor with valid parameters');

</script>
